Remove separate registry package
