+++ .github/CODEOWNERS
@@ -120,7 +120,6 @@
 /pkg/services/queryhistory/ @grafana/backend-platform
 /pkg/services/quota/ @grafana/backend-platform
 /pkg/services/rendering/ @grafana/backend-platform
-/pkg/services/registryentity/ @grafana/backend-platform
 /pkg/services/screenshot/ @grafana/backend-platform
 /pkg/services/search/ @grafana/backend-platform
 /pkg/services/searchusers/ @grafana/backend-platform
