--- pkg/services/registryentity/model.go
@@ -1,5 +1,0 @@
-package registryentity
-
-import "errors"
-
-var ErrTargetSrvConflict = errors.New("target srv conflict")
